+++ packages/transport-chrome/src/test-utils/unhandled.ts
@@ -1,19 +1,5 @@
 /* eslint-disable @typescript-eslint/no-non-null-assertion -- test utils */
-import { expect, MockedFunction, vi } from 'vitest';
-
-/**
- * Returns the last result of a mocked function.
- */
-// eslint-disable-next-line @typescript-eslint/no-explicit-any -- test utility
-export const lastResult = <T extends (...args: any[]) => any>(
-  mfn: MockedFunction<T>,
-): ReturnType<T> => {
-  const last = mfn.mock.results.at(-1);
-  if (last?.type === 'return') {
-    return last.value as never;
-  }
-  throw last?.value;
-};
+import { expect, vi } from 'vitest';
 
 /** Default for replaced listeners. */
 const replacedUnhandledExceptionLogger: NodeJS.UncaughtExceptionListener = (
